move data classes out + single client
